Move functions related to thread into ThreadHelper.
